+++ examples/imagenet/train.py
@@ -27,7 +27,7 @@ from clu import metric_writers
 from clu import periodic_actions
 import flax
 from flax import jax_utils
-from flax import optim
+from flax.optim import dynamic_scale as dynamic_scale_lib
 from flax.training import checkpoints
 from flax.training import common_utils
 from flax.training import train_state
@@ -191,7 +191,7 @@ def create_input_iter(dataset_builder, batch_size, image_size, dtype, train,
 
 class TrainState(train_state.TrainState):
   batch_stats: Any
-  dynamic_scale: flax.optim.DynamicScale
+  dynamic_scale: dynamic_scale_lib.DynamicScale
 
 
 def restore_checkpoint(state, workdir):
@@ -224,7 +224,7 @@ def create_train_state(rng, config: ml_collections.ConfigDict,
   dynamic_scale = None
   platform = jax.local_devices()[0].platform
   if config.half_precision and platform == 'gpu':
-    dynamic_scale = optim.DynamicScale()
+    dynamic_scale = dynamic_scale_lib.DynamicScale()
   else:
     dynamic_scale = None