+++ flax/linen/combinators.py
@@ -1,8 +1,23 @@
+# Copyright 2022 The Flax Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 """Combinators of modules, such as a Sequential."""
 
-from typing import Callable, Sequence
+from typing import Any, Callable, Sequence
 
 from flax.linen.module import Module
+
 
 class Sequential(Module):
   """Applies a linear chain of Modules.
@@ -28,7 +43,7 @@ class Sequential(Module):
                               nn.Dense(2),
                               nn.log_softmax])(x)
   """
-  layers: Sequence[Callable]
+  layers: Sequence[Callable[..., Any]]
 
   def __call__(self, *args, **kwargs):
     if not self.layers: